files/VerifAI_Complete_Codebase/verifai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.28
langchain-anthropic==0.2.4
langchain-community==0.3.1
langchain-core==0.3.8
anthropic==0.34.2
tavily-python==0.5.0
pillow==10.4.0
exifread==3.0.0
librosa==0.10.2
numpy==1.26.4
httpx==0.27.2
python-multipart==0.0.9
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
